+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - **EXPERIMENTAL_Modal** `ref` not properly passed to modal element.
+- **EXPERIMENTAL_Table** `emptyState` height
 
 ## [9.133.1] - 2020-10-29